Fixed hardcoded values to global constants

+++ Squids-Movies-App/SquidsMovieApp.Core/Factories/Contracts/IUserFactory.cs
@@ -10,6 +10,6 @@ namespace SquidsMovieApp.Core.Factories.Contracts
     public interface IUserFactory
     {
         UserModel CreateUserModel(string firstName, string lastName, int? age, string nickName,
-            string email, string password, bool isAdmin, int moneyBalance);
+            string email, string password, bool isAdmin, decimal moneyBalance);
     }
 }
